Update README.md
The difference between the data structures Vectors, Lists and Data Frame.
A vector is a group of data elements that are of the same data type. 
A List is also a group of data elements that consists of different data types.
A data frame is a collection of columns that are similar to a spreadsheet table.